pyproject.toml
[project]
name = "flowreconstructionfromexperiment"
version = "0.1.0"
description = "Recontruct flows from sparse sensor measurements."
readme = "README.md"
requires-python = ">=3.11.0,<3.12.0"
dependencies = [
    "chex==0.1.86",
    "dm-haiku==0.0.13",
    "flax>=0.8.4",
    "h5py==3.9.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "ml-collections==0.1.1",
    "nbformat>=5.10.4",
    "numpy<2.0.0",
    "optax==0.1.7",
    "pip>=25.0.1",
    "plotly>=6.0.1",
    "pyyaml==6.0",
    "scipy==1.12.0",
    "tqdm>=4.67.1",
    "wandb>=0.15.4",
]


[project.optional-dependencies]
cpu = ["jax[cpu]==0.4.24",]
gpu = [
 	"jax[cuda12]==0.4.24",
    "nvidia-cudnn-cu12<9.0.0.0",
]
